requirements.txt
streamlit
python-dotenv
openai
pdfplumber
pytesseract
Pillow
chromadb
crewai
tiktoken
pydantic
